Fixed CCR scrolling issues

+++ include/command_macros_base.h
@@ -12,7 +12,7 @@
 #else
 #define CMD_BBBB(a, b, c, d) (_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 8, 8) | _SHIFTL(d, 0, 8))
 #define CMD_BBH(a, b, c) (_SHIFTL(a, 24, 8) | _SHIFTL(b, 16, 8) | _SHIFTL(c, 0, 16))
-#define CMD_HH(a, b) (_SHIFTL(a, 16, 16) | _SHIFTL(b, 0, 16))
+#define CMD_HH(a, b) (_SHIFTL((s16)a, 16, 16) | _SHIFTL((s16)b, 0, 16))
 #define CMD_W(a) (a)
 #endif
 #else
